Refactored Flask app for containerization #25
