+++ environment.yml
@@ -3,122 +3,125 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - appdirs=1.4.4=pyh9f0ad1d_0
-  - arm_pyart=1.11.6=py38h6f4d8f0_1
-  - beautifulsoup4=4.10.0=pyha770c72_0
-  - bidict=0.21.2=pyhd3eb1b0_0
-  - blas=1.0=mkl
-  - boto=2.49.0=py38_0
-  - brotlipy=0.7.0=py38h2bbff1b_1003
-  - bzip2=1.0.8=h8ffe710_4
-  - ca-certificates=2021.10.8=h5b45459_0
-  - cartopy=0.18.0=py38h80a4efb_1
-  - certifi=2021.10.8=py38haa244fe_1
-  - cffi=1.15.0=py38h2bbff1b_0
-  - cftime=1.5.2=py38h6f4d8f0_0
-  - charset-normalizer=2.0.10=pyhd8ed1ab_0
-  - click=8.0.3=pyhd3eb1b0_0
-  - cryptography=36.0.0=py38h21b164f_0
-  - curl=7.80.0=h2bbff1b_0
-  - cycler=0.11.0=pyhd3eb1b0_0
-  - dataclasses=0.8=pyh6d0b6a4_7
-  - dnspython=2.2.0=pyhd8ed1ab_0
-  - eventlet=0.33.0=pyhd8ed1ab_0
-  - flask=1.1.2=pyhd3eb1b0_0
-  - flask-cors=3.0.10=pyhd3eb1b0_0
-  - flask-socketio=5.1.0=pyhd3eb1b0_0
-  - freetype=2.10.4=hd328e21_0
-  - geos=3.8.0=h33f27b4_0
-  - gevent=21.1.2=py38h2bbff1b_1
-  - greenlet=1.1.1=py38hd77b12b_0
-  - hdf4=4.2.15=h0e5069d_3
-  - hdf5=1.12.1=nompi_h2a0e4a3_100
-  - icc_rt=2019.0.0=h0cc432a_1
-  - icu=58.2=ha925a31_3
-  - idna=3.3=pyhd3eb1b0_0
-  - imageio=2.6.1=py38_0
-  - importlib-metadata=4.10.1=py38haa244fe_0
-  - importlib_metadata=4.10.1=hd8ed1ab_0
-  - importlib_resources=5.4.0=pyhd8ed1ab_0
-  - intel-openmp=2021.4.0=haa95532_3556
-  - itsdangerous=2.0.1=pyhd3eb1b0_0
-  - jinja2=3.0.2=pyhd3eb1b0_0
-  - jpeg=9d=h2bbff1b_0
-  - kafka-python=2.0.2=pyh9f0ad1d_0
-  - kiwisolver=1.3.1=py38hd77b12b_0
-  - libcurl=7.80.0=h86230a5_0
-  - libnetcdf=4.8.1=nompi_h1cc8e9d_101
-  - libpng=1.6.37=h2a8f88b_0
-  - libprotobuf=3.19.1=h23ce68f_0
-  - libssh2=1.10.0=h680486a_2
-  - libtiff=4.2.0=hd0e1b90_0
-  - libwebp=1.2.0=h2bbff1b_0
-  - libzip=1.8.0=hfed4ece_1
-  - lz4-c=1.9.3=h2bbff1b_1
-  - markupsafe=2.0.1=py38h2bbff1b_0
-  - matplotlib=3.3.4=py38haa95532_0
-  - matplotlib-base=3.3.4=py38h49ac443_0
-  - metpy=1.2.0=pyhd8ed1ab_0
-  - mkl=2021.4.0=haa95532_640
-  - mkl-service=2.4.0=py38h2bbff1b_0
-  - mkl_fft=1.3.1=py38h277e83a_0
-  - mkl_random=1.2.2=py38hf11a4ad_0
-  - netcdf4=1.5.8=nompi_py38h0500770_101
-  - numpy=1.21.2=py38hfca59bb_0
-  - numpy-base=1.21.2=py38h0829f74_0
-  - olefile=0.46=pyhd3eb1b0_0
-  - openssl=1.1.1l=h8ffe710_0
-  - packaging=21.3=pyhd8ed1ab_0
-  - pandas=1.2.4=py38hd77b12b_0
-  - pillow=8.2.0=py38h4fa10fc_0
-  - pint=0.18=pyhd8ed1ab_0
-  - pip=21.2.2=py38haa95532_0
-  - pooch=1.6.0=pyhd8ed1ab_0
-  - proj=6.2.1=h9f7ef89_0
-  - protobuf=3.19.1=py38hd77b12b_0
-  - pycparser=2.21=pyhd3eb1b0_0
-  - pyopenssl=21.0.0=pyhd3eb1b0_1
-  - pyparsing=3.0.4=pyhd3eb1b0_0
-  - pyproj=2.6.1.post1=py38h593ac45_1
-  - pyqt=5.9.2=py38ha925a31_4
-  - pyshp=2.1.3=pyh44b312d_0
-  - pysocks=1.7.1=py38haa95532_0
-  - python=3.8.8=hdbf39b2_5
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - python-engineio=4.1.0=pyhd3eb1b0_0
-  - python-socketio=5.3.0=pyhd3eb1b0_0
-  - python_abi=3.8=2_cp38
-  - pytz=2021.3=pyhd3eb1b0_0
-  - qt=5.9.7=vc14h73c81de_0
-  - requests=2.27.1=pyhd8ed1ab_0
-  - scipy=1.7.3=py38h0a974cb_0
-  - setuptools=52.0.0=py38haa95532_1
-  - shapely=1.7.1=py38h06580b3_0
-  - sip=4.19.13=py38ha925a31_0
-  - siphon=0.9=pyhd8ed1ab_2
-  - six=1.16.0=pyhd3eb1b0_0
-  - soupsieve=2.3.1=pyhd8ed1ab_0
-  - sqlite=3.35.4=h2bbff1b_0
-  - tk=8.6.11=h2bbff1b_0
-  - tornado=6.1=py38h2bbff1b_0
-  - traitlets=5.1.1=pyhd8ed1ab_0
-  - typing_extensions=4.0.1=pyha770c72_0
-  - urllib3=1.26.4=pyhd3eb1b0_0
-  - vc=14.2=h21ff451_1
-  - vs2015_runtime=14.27.29016=h5e58377_2
-  - werkzeug=2.0.2=pyhd3eb1b0_0
-  - wheel=0.37.1=pyhd3eb1b0_0
-  - win_inet_pton=1.1.0=py38haa95532_0
-  - wincertstore=0.2=py38haa95532_2
-  - xarray=0.20.2=pyhd8ed1ab_0
-  - xlsxwriter=1.3.8=pyhd3eb1b0_0
-  - xz=5.2.5=h62dcd97_0
-  - zipp=3.7.0=pyhd8ed1ab_0
-  - zlib=1.2.11=h8cc25b3_4
-  - zope=1.0=py38_1
-  - zope.event=4.5.0=py38_0
-  - zope.interface=5.4.0=py38h2bbff1b_0
-  - zstd=1.4.9=h19a0ad4_0
+  - appdirs=1.4.4
+  - arm_pyart=1.11.6
+  - beautifulsoup4=4.10.0
+  - bidict=0.21.2
+  - blas=1.0
+  - boto=2.49.0
+  - brotlipy=0.7.0
+  - bzip2=1.0.8
+  - ca-certificates=2021.10.26
+  - cartopy=0.18.0
+  - certifi=2021.10.8
+  - cffi=1.15.0
+  - cftime=1.5.2
+  - charset-normalizer=2.0.10
+  - click=8.0.3
+  - colorama=0.4.4
+  - cryptography=36.0.0
+  - curl=7.80.0
+  - cycler=0.11.0
+  - dataclasses=0.8
+  - dnspython=2.2.0
+  - eventlet=0.33.0
+  - flask=1.1.2
+  - flask-cors=3.0.10
+  - flask-socketio=5.1.0
+  - freetype=2.10.4
+  - geos=3.8.0
+  - gevent=21.1.2
+  - greenlet=1.1.1
+  - hdf4=4.2.15
+  - hdf5=1.12.1
+  #- icc_rt=2019.0.0
+  - icu=58.2
+  - idna=3.3
+  - imageio=2.6.1
+  - importlib-metadata=4.10.1
+  - importlib_metadata=4.10.1
+  - importlib_resources=5.4.0
+  - intel-openmp=2021.4.0
+  - itsdangerous=2.0.1
+  - jinja2=3.0.2
+  - jpeg=9d
+  - kafka-python=2.0.2
+  - kiwisolver=1.3.1
+  - libcurl=7.80.0
+  - libnetcdf=4.8.1
+  - libpng=1.6.37
+  - libprotobuf=3.19.1
+  - libssh2=1.10.0
+  - libtiff=4.2.0
+  - libwebp=1.2.0
+  - libzip=1.8.0
+  - lz4-c=1.9.3
+  - markupsafe=2.0.1
+  - matplotlib=3.3.4
+  - matplotlib-base=3.3.4
+  - metpy=1.2.0
+  - mkl=2021.4.0
+  - mkl-service=2.4.0
+  - mkl_fft=1.3.1
+  - mkl_random=1.2.2
+  - netcdf4=1.5.8
+  - numpy=1.21.2
+  - numpy-base=1.21.2
+  - olefile=0.46
+  - openssl=1.1.1m
+  - packaging=21.3
+  - pandas=1.2.4
+  - pillow=8.2.0
+  - pint=0.18
+  - pip=21.2.2
+  - pooch=1.6.0
+  - proj=6.2.1
+  - protobuf=3.19.1
+  - pycparser=2.21
+  - pyopenssl=21.0.0
+  - pyparsing=3.0.4
+  - pyproj=2.6.1.post1
+  - pyqt=5.9.2
+  - pyshp=2.1.3
+  - pysocks=1.7.1
+  - python=3.8.8
+  - python-dateutil=2.8.2
+  - python-dotenv=0.19.2
+  - python-engineio=4.1.0
+  - python-socketio=5.3.0
+  - python_abi=3.8
+  - pytz=2021.3
+  - qt=5.9.7
+  - requests=2.27.1
+  - scipy=1.7.3
+  - setuptools=52.0.0
+  - shapely=1.7.1
+  - sip=4.19.13
+  - siphon=0.9
+  - six=1.16.0
+  - soupsieve=2.3.1
+  - sqlite=3.35.4
+  - tk=8.6.11
+  - tornado=6.1
+  - traitlets=5.1.1
+  - typing_extensions=4.0.1
+  - urllib3=1.26.4
+  #- vc=14.2
+  #- vs2015_runtime=14.27.29016
+  - werkzeug=2.0.2
+  - wheel=0.37.1
+  #- win_inet_pton=1.1.0
+  #- wincertstore=0.2
+  - xarray=0.20.2
+  - xlsxwriter=1.3.8
+  - xz=5.2.5
+  - zipp=3.7.0
+  - zlib=1.2.11
+  - zope=1.0
+  - zope.event=4.5.0
+  - zope.interface=5.4.0
+  - zstd=1.4.9
   - pip:
+    - flask-kafka==0.0.3
     - gunicorn==20.1.0
 prefix: C:\Users\vikra\anaconda3\envs\env
