--- startup.bat
@@ -1,3 +1,0 @@
-set FLASK_APP=flask-producer-app
-set FLASK_ENV=development
-flask run --host=localhost --port=5000
