+++ weather_reporter/config.env
@@ -1,0 +1,6 @@
+FLASK_HOST = '0.0.0.0'
+FLASK_PORT = '5000'
+KAFKA_SERVER = 'localhost:9092'
+REQUEST_TOPIC = 'REQUEST'
+RESPONSE_TOPIC = 'WEATHER_REPORT'
+GROUP_ID = 'test-consumer-group'
